Move logic from lib/main.js to lib/test-worker.js
This focuses lib/main.js on creating and exporting the runner, while the
worker focuses on managing IPC, catching errors, and doing generally
doing the work.

The worker now guards against repeatedly sending stats, or sending test
results after it's started to exit.

Required and test files are loaded at the end, after handlers for
uncaught exceptions and unhandled rejections have been installed.
