+++ lib/test-worker.js
@@ -19,32 +19,35 @@
 const currentlyUnhandled = require('currently-unhandled')();
 const isObj = require('is-obj');
 
+const nowAndTimers = require('./now-and-timers');
 const adapter = require('./process-adapter');
 const serializeError = require('./serialize-error');
 const opts = require('./worker-options').get();
 
-// Install before processing opts.require, so if helpers are added to the
-// require configuration the *compiled* helper will be loaded.
-adapter.installSourceMapSupport();
-adapter.installPrecompilerHook();
-
-(opts.require || []).forEach(x => {
-	if (/[/\\]@std[/\\]esm[/\\]index\.js$/.test(x)) {
-		require = require(x)(module); // eslint-disable-line no-global-assign
-	} else {
-		require(x);
-	}
-});
-
-const testPath = opts.file;
-
+// Store details about the test run, to be sent to the parent process later.
 const dependencies = new Set();
-adapter.installDependencyTracking(dependencies, testPath);
-
 const touchedFiles = new Set();
 
 // Set when main.js is required (since test files should have `require('ava')`).
 let runner = null;
+
+// Track when exiting begins, to avoid repeatedly sending stats, or sending
+// individual test results once stats have been sent. This is necessary since
+// exit() can be invoked from the worker process and over IPC.
+let exiting = false;
+function exit() {
+	if (exiting) {
+		return;
+	}
+	exiting = true;
+
+	// Reference the IPC channel so the exit sequence can be completed.
+	adapter.ipcChannel.ref();
+
+	const stats = runner.buildStats();
+	adapter.send('results', {stats});
+}
+
 exports.setRunner = newRunner => {
 	runner = newRunner;
 	runner.on('dependency', file => {
@@ -55,14 +58,26 @@ exports.setRunner = newRunner => {
 			touchedFiles.add(file);
 		}
 	});
+	runner.on('test', props => {
+		if (exiting) {
+			return;
+		}
+
+		const hasError = typeof props.error !== 'undefined';
+		// Don't send anything if it's a passed hook
+		if (!hasError && props.type !== 'test') {
+			return;
+		}
+
+		if (hasError) {
+			props.error = serializeError(props.error);
+		} else {
+			props.error = null;
+		}
+
+		adapter.send('test', props);
+	});
 };
-
-require(testPath);
-
-// If AVA was not required, show an error
-if (!runner) {
-	adapter.send('no-tests', {avaRequired: false});
-}
 
 function attributeLeakedError(err) {
 	if (!runner) {
@@ -72,14 +87,7 @@ function attributeLeakedError(err) {
 	return runner.attributeLeakedError(err);
 }
 
-const attributedRejections = new Set();
-process.on('unhandledRejection', (reason, promise) => {
-	if (attributeLeakedError(reason)) {
-		attributedRejections.add(promise);
-	}
-});
-
-process.on('uncaughtException', exception => {
+function handleUncaughtException(exception) {
 	if (attributeLeakedError(exception)) {
 		return;
 	}
@@ -98,12 +106,21 @@ process.on('uncaughtException', exception => {
 		};
 	}
 
-	// Ensure the IPC channel is refereced. The uncaught exception will kick off
+	// Ensure the IPC channel is referenced. The uncaught exception will kick off
 	// the teardown sequence, for which the messages must be received.
 	adapter.ipcChannel.ref();
 
 	adapter.send('uncaughtException', {exception: serialized});
+}
+
+const attributedRejections = new Set();
+process.on('unhandledRejection', (reason, promise) => {
+	if (attributeLeakedError(reason)) {
+		attributedRejections.add(promise);
+	}
 });
+
+process.on('uncaughtException', handleUncaughtException);
 
 let tearingDown = false;
 process.on('ava-teardown', () => {
@@ -112,6 +129,9 @@ process.on('ava-teardown', () => {
 		return;
 	}
 	tearingDown = true;
+
+	// Reference the IPC channel so the teardown sequence can be completed.
+	adapter.ipcChannel.ref();
 
 	let rejections = currentlyUnhandled()
 		.filter(rejection => !attributedRejections.has(rejection.promise));
@@ -143,3 +163,71 @@ process.on('ava-teardown', () => {
 process.on('ava-exit', () => {
 	process.exit(0); // eslint-disable-line xo/no-process-exit
 });
+
+process.on('ava-init-exit', () => {
+	exit();
+});
+
+// Store value in case to prevent required modules from modifying it.
+const testPath = opts.file;
+
+// Install before processing opts.require, so if helpers are added to the
+// require configuration the *compiled* helper will be loaded.
+adapter.installDependencyTracking(dependencies, testPath);
+adapter.installSourceMapSupport();
+adapter.installPrecompilerHook();
+
+try {
+	(opts.require || []).forEach(x => {
+		if (/[/\\]@std[/\\]esm[/\\]index\.js$/.test(x)) {
+			require = require(x)(module); // eslint-disable-line no-global-assign
+		} else {
+			require(x);
+		}
+	});
+
+	require(testPath);
+} catch (err) {
+	handleUncaughtException(err);
+} finally {
+	if (runner) {
+		nowAndTimers.setImmediate(() => {
+			const hasExclusive = runner.tests.hasExclusive;
+			const numberOfTests = runner.tests.testCount;
+
+			if (numberOfTests === 0) {
+				adapter.send('no-tests', {avaRequired: true});
+				return;
+			}
+
+			adapter.send('stats', {
+				testCount: numberOfTests,
+				hasExclusive
+			});
+
+			process.on('ava-run', options => {
+				// Unreference the IPC channel. This stops it from keeping the event loop
+				// busy, which means the `beforeExit` event can be used to detect when tests
+				// stall.
+				adapter.ipcChannel.unref();
+
+				runner.run(options)
+					.then(() => {
+						runner.saveSnapshotState();
+
+						return exit();
+					})
+					.catch(err => {
+						handleUncaughtException(err);
+					});
+			});
+
+			process.on('ava-init-exit', () => {
+				exit();
+			});
+		});
+	} else {
+		// If AVA was not required, show an error
+		adapter.send('no-tests', {avaRequired: false});
+	}
+}
